% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENN.R
\name{ENN}
\alias{ENN}
\title{Perform Edited Nearest Neighbors (ENN) downsampling}
\arguments{
\item{data}{Data frame containing the dataset.
It should include columns for predictors (features) and a response variable.
The response variable should be a factor with two levels.}
}
\value{
A data frame containing the downsampled majority class instance
        satisfying the ENN criteria.
}
\description{
This function performs Edited Nearest Neighbors (ENN) downsampling
for the majority class instances in the dataset.
}
\examples{

# Recommended - Apply Parallel Processing
library(doParallel)
library(dplyr)
library(magrittr)
library(foreach)
library(knitr)
library(tidyr)
#registerDoParallel(detectCores())
#Alternative max level of cores, but need to manually set on certain machines due to restrictions
registerDoParallel(2)

#load data
data(iris)

# rename df
df <- iris

#make setosa binary
df$setosaYN <- as.numeric(df$Species == "setosa")

# Remove the "Species" column
df <- df[, -5]

# Remove the "Species" column and move "setosaYN" to the beginning
df <- df[, c("setosaYN", setdiff(names(df), c("Species", "setosaYN")))]

# Preprocess data into specified df
X <- df[, -1]
y <- df[, 1]
data <- data.frame(X, y)

# Apply ENN Majority Class Down Sampling and output df
data2 <- ENN(data)

# notice that it did not detect any samples to be selected for down sampling

}
