% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enn_downsample_instance.R
\name{enn_downsample_instance}
\alias{enn_downsample_instance}
\title{Perform Edited Nearest Neighbors (ENN) downsampling for a single instance}
\arguments{
\item{i}{Index of the instance in the dataset.}

\item{data}{Data frame containing the dataset.}

\item{k}{Number of nearest neighbors to consider.}
}
\value{
A data frame containing the instance if it satisfies the ENN criteria, otherwise NULL.
}
\description{
This function performs Edited Nearest Neighbors (ENN) downsampling for a single instance.
If the majority class instance is misclassified by its neighbors, it is excluded.
}
\keyword{internal}
