% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean_distance.R
\name{euclidean_distance}
\alias{euclidean_distance}
\title{Calculate Euclidean distance between two points}
\usage{
euclidean_distance(a, b)
}
\arguments{
\item{a}{Numeric vector representing the first point.}

\item{b}{Numeric vector representing the second point.}
}
\value{
The Euclidean distance between points \code{a} and \code{b}.
}
\description{
This function calculates the Euclidean distance between two points.
}
\keyword{internal}
