% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_neighbors.R
\name{find_neighbors}
\alias{find_neighbors}
\title{Find k nearest neighbors for a given instance}
\arguments{
\item{data}{Data frame containing the dataset.}

\item{instance}{Numeric vector representing the instance for which neighbors are sought.}

\item{k}{Number of neighbors to find.}
}
\value{
A vector of indices corresponding to the k nearest neighbors.
}
\description{
This function finds the indices of the k nearest neighbors for a given instance
in a dataset using Euclidean distance.
}
\keyword{internal}
